Reduce delay time and enhance message sending
Adjusted the delay time in TestService.java from 10000 to 5000 to shorten message latency. Also, made changes to RabbitService.java, including the add of PERSISTENT delivery mode and an expiration time of 10000 for the messages. These changes improve the efficiency of the messaging service by reducing the delay time and ensuring messages are not lost if the broker restarts. In RabbitmqDelayQueueApplication.java, Beans for DirectExchange and Binding were added to manage the RabbitMQ delayed queues creating a more robust and efficient service.